各UIのstruct情報が欲しかったためManagerを変えた

+++ アゲアゲバード/Scene/GameEndScene.cpp
@@ -43,8 +43,8 @@ void GameEndScene::InitScene()
 	{
 		LoadTexture("Res/Tex/ƒNƒŠƒA.png", &m_GameEndSceneInfo.m_end);
 	}
-	m_ui_manager->Init((int)UIManager::Scene::end);
-	m_ui_manager->LoadFile((int)UIManager::Scene::end);
+	m_ui_manager->Init(UIManager::Scene::end);
+	m_ui_manager->LoadFile(UIManager::Scene::end);
 	m_sound_manager->RegisterEndSound();
 	m_sound_manager->SoundBGM(SceneManager::Instance()->GetOptionScene()->GetOptionSceneInfo()->m_sound_volume);
 	m_scene_manager->SetSceneInfo()->m_CurrentSceneStep=SceneStep::MainStep;
